backend/requirements.txt
openai
supabase
python-dotenv
requests